Moves data files
